requirements.txt
sounddevice==0.4.6
numpy==1.26.4
websockets==12.0
httpx==0.27.0
keyboard==0.13.5
pyperclip==1.8.2
pyautogui==0.9.54
Pillow==10.3.0
keyring==25.2.1
pycaw==20240210
comtypes==1.4.1
pywin32==311
